Gearbox in Input

+++ core/src/main/java/tcs/racer_game/car/Gearbox.java
@@ -3,11 +3,11 @@ package tcs.racer_game.car;
 public interface Gearbox {
 
     ///  value is in percent (%)
-    void setBrake(State state, int value);
+    void setBrake(State state, float value);
 
-    void setGas(State state, int value);
+    void setGas(State state, float value);
 
-    void setTurn(State state, Direction direction, int value);
+    void setTurn(State state, Direction direction, float value);
 
 
 }
