+++ core/src/main/java/tcs/racer_game/car/carv2/simplecar/CarV2Gearbox.java
@@ -7,16 +7,16 @@ import tcs.racer_game.car.State;
 public class CarV2Gearbox implements Gearbox {
 
     State stateBrake = State.OFF;
-    int valueBrake = 0;
+    float valueBrake = 0;
     State stateGas = State.OFF;
-    int valueGas = 0;
+    float valueGas = 0;
     State stateTurn = State.OFF;
     Direction turnDirection = Direction.none;
-    int valueTurn = 0;
+    float valueTurn = 0;
 
 
     @Override
-    public void setBrake(State state, int value) {
+    public void setBrake(State state, float value) {
         if(value < 0 || value > 100){
             throw new IllegalArgumentException("value lesser then zero in Gearbox");
         }
@@ -29,7 +29,7 @@ public class CarV2Gearbox implements Gearbox {
     }
 
     @Override
-    public void setGas(State state, int value) {
+    public void setGas(State state, float value) {
         if(value < 0 || value > 100){
             throw new IllegalArgumentException("value lesser then zero in Gearbox");
         }
@@ -42,7 +42,7 @@ public class CarV2Gearbox implements Gearbox {
     }
 
     @Override
-    public void setTurn(State state, Direction direction, int value) {
+    public void setTurn(State state, Direction direction, float value) {
         if(value < 0){
             throw new IllegalArgumentException("value lesser then zero in Gearbox");
         }
